THE CODE/pyproject.toml
[project]
name = "neuro-backend"
version = "0.1.0"
description = "Neuropetitorius AI Tutor Backend MVP"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "google-genai>=0.3.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
    "structlog>=24.4.0",
    "python-multipart>=0.0.17",
    "httpx>=0.28.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]